fix(kffi): correct engine doc, fixture ABI and null-fn guard

+++ kffi/src/ffiTest/resources/downcall_fixture.c
@@ -2,8 +2,8 @@
 
 uint64_t bench_empty(void) { return 42u; }
 
-uint64_t bench_add4(uint64_t a, uint64_t b, uint64_t c, uint64_t d) {
-    return a + b + c + d;
+int64_t bench_add4(int32_t a, int32_t b, int32_t c, int32_t d) {
+    return (int64_t)a + b + c + d;
 }
 
 uint64_t bench_add8(uint64_t a, uint64_t b, uint64_t c, uint64_t d,
