+++ kffi/src/ffiTest/resources/downcall_fixture.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 
 uint64_t bench_empty(void);
-uint64_t bench_add4(uint64_t a, uint64_t b, uint64_t c, uint64_t d);
+int64_t bench_add4(int32_t a, int32_t b, int32_t c, int32_t d);
 uint64_t bench_add8(uint64_t a, uint64_t b, uint64_t c, uint64_t d,
                     uint64_t e, uint64_t f, uint64_t g, uint64_t h);
 void *bench_roundtrip_ptr(void *p);
